Fix introduction advice on generated types with error checking APIs (#11654)
The previous fix threw a postpone exception when accessing interfaces but this creates a catch 22 since type element visitors need to be able to navigate the AST without errors.

However it a bean definition should never be generated if the type has errors in the signature. So this adds special handling for BeanDefinitionInjectProcessor to postpone if there are errors and it hasn't already been postponed.

Possibly we should hard fail and adds APIs to retrieve the error.