When checking `refCnt > 0`, make sure we're the exclusive owner beforehand (#11864)
When checking `refCnt > 0`, make sure we're the exclusive owner beforehand
In `dataStream`, we still use the antipattern of a conditional release after onNext. The idea is that if the downstream consumer does not release the buffer, we'll do it for them after onNext.

However, if the downstream *does* release the buffer but the buffer is also still used elsewhere, this `refCnt` check did not work, and the buffer was released twice. This issue likely was exposed by the new netty allocator.

This patch makes sure the buffer refCnt is 1 so that we don't interfere with others.

Fixes https://github.com/micronaut-projects/micronaut-reactor/issues/555