Fix StreamingFileUpload leak (#11898)
Ensure that the NettyStreamingFileUpload is discarded after transfer.

The test is in the TCK. I also added leak detection infrastructure to the TCK, and moved the http-server-netty leak detection to use the same code.

There was also a minor leak in the http client uncovered by enabling detection in the TCK.

Fixes #11897