fix: Allow other executor instrumentations than context propagation (#11863)
fixes #11653

The `InstrumentedExecutorService` interface is also used in micrometer.

The current instrumentation only checks if a given executor service is instrumented, but there is no notion of which specific instrumentation it is instrumented with.

This means that if the [micrometer instrumentation](https://github.com/micronaut-projects/micronaut-micrometer/blob/56217245d1cc33cf6894c925ebed71eaf990dc28/micrometer-core/src/main/java/io/micronaut/configuration/metrics/binder/executor/ExecutorServiceMetricsBinder.java#L91-L125) is applied first, the context propagation instrumentation is skipped.

To solve this for context propagation, we provide an implementation of the `InstrumentedExecutorService` and traverse the targets to find that specific implementation. It's not perfect, but should be better than before.