4.7.x -> 4.8.x (#12028)
* Update settings.gradle

* [skip ci] Release v4.7.26

* chore: Bump version to 4.7.27-SNAPSHOT

* Backport three fixes to 4.7.x and fix logged exception in response handling (#11921)

Properly discard inbound HttpContent when connection cancels before readComplete (#11496)

When a connection is closed before readComplete is called, OptimisticBufferingInboundHandler would not release the input buffer correctly.

Discovered by fuzzing with EmbeddedChannel. I'm not sure if this can happen with a non-embedded channel, but better safe than sorry. A test is not feasible, detection relies on finding dangling buffers, which the test suite is too noisy for.

* Fix buffer leak when decompressing empty LastHttpContent (#12011) (#12013)

* Fix buffer leak when decompressing empty LastHttpContent
Discovered by fuzzing

* Update netty (#12026)

* Fix buffer leak when decompressing empty LastHttpContent (#12011)

* Fix buffer leak when decompressing empty LastHttpContent
Discovered by fuzzing

* fix

* update netty

* [skip ci] Release v4.7.27

---------

Co-authored-by: Graeme Rocher <graeme.rocher@oracle.com>
Co-authored-by: micronaut-build <micronaut-build-account@grails.org>