Event loop carrier support for openjdk loom branch (#12061)
Add support for the APIs exposed by the [openjdk loom branch](https://github.com/openjdk/loom/). I kept it as close as possible to the current code, haven't checked for optimization potential yet.

The reflection code is particularly messy this time. I had to implement the new VirtualThreadScheduler interface, and that required messing with LambdaMetafactory.