Annotations referencing generated sources result in PostponeToNextRoundException (#12110)
Currently if you have an annotation that references a type generated by a source generator this causes `PostponeToNextRoundException` to be thrown by the compiler. 

There is a codepath that processes introduction advice annotations that does not handle `PostponeToNextRoundException`. This PR adjusts that code path to handle this exception and postpone the introduction advice to be handled in the next round of processing.

---------

Co-authored-by: Denis Stepanov <denis.s.stepanov@oracle.com>