Fix test-suite-kotlin-graalvm
This project setup is a bit special: it uses the test library
convention, which brings Groovy and Spock on classpath.
However, these are not supported with GraalVM, which causes
a compilation failure.

The workaround, and safest change until we understand better
the project setup and if it really requires the base plugin,
is to exclude groovy.